backend/requirements.txt
# ============================
# Core Web Framework
# ============================
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# ============================
# Data validation
# ============================
pydantic>=2.5,<3
pydantic-settings>=2.1

# ============================
# LLM / AI
# ============================
groq>=0.5.0
sentence-transformers>=2.3.1
qdrant-client>=1.7.0

# ============================
# Database
# ============================
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9.9
redis>=5.0.1

# ============================
# Auth & security
# ============================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-multipart>=0.0.6


# ============================
# Content extraction
# ============================
beautifulsoup4>=4.12.3
lxml>=5.1.0
html2text>=2024.2.26
readability-lxml>=0.8.1

# ============================
# Utilities
# ============================
python-dotenv>=1.0.0
httpx>=0.26.0
tenacity>=8.2.3

# ============================
# Monitoring
# ============================
prometheus-client>=0.19.0
sentry-sdk[fastapi]>=1.39.2

# ============================
# Testing
# ============================
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0

# ============================
# Development tools
# ============================
black>=24.1.1
flake8>=7.0.0
mypy>=1.8.0

asyncpg==0.29.0
google-auth
requests
